Finished with login

+++ model/contacts.js
@@ -1,12 +1,18 @@
 const Contact = require('./schemas/contact')
 
 const listContacts = async (userId) => {
-  const results = await Contact.find({ owner: userId })
+  const results = await Contact.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription -_id',
+  })
   return results
 }
 
-const getContactById = async (contactId) => {
-  const result = await Contact.findOne({ _id: contactId })
+const getContactById = async (contactId, userId) => {
+  const result = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: 'owner',
+    select: 'email subscription -_id',
+  })
   return result
 }
 
@@ -15,17 +21,17 @@ const addContact = async (body) => {
   return result
 }
 
-const updateContact = async (id, body) => {
+const updateContact = async (id, body, userId) => {
    const result = await Contact.findByIdAndUpdate(
-    { _id: id },
+    { _id: id, owner: userId },
     { ...body },
     { new: true }
   )
   return result 
 }
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findByIdAndRemove({ _id: contactId })
+const removeContact = async (contactId, userId) => {
+  const result = await Contact.findByIdAndRemove({ _id: contactId, owner: userId })
   return result 
 }
